requirements.txt
discord.py
pymongo
dnspython
dex.py
EZPaginator